updated the event files
updated the event repository added new methods. also updated the event services, changed the namespace and added new methods
